test with streams

+++ src/cuda/gemm_tensorcores/tst/tensor_dmr/simpleTensorCoreGEMM.cu
@@ -524,10 +524,10 @@ int main(int argc, char* argv[]) {
 
   
   for (int i = 0; i <  20; i++) {      
-    float v1 = d_host_wmma[i];
-    float v2 = d_host_sw[i];
+    half v1 = d_host_wmma[i];
+    half v2 = d_host_sw[i];
     //float v3 = d_host_cublas[i]; 
-    float v4 = fabs(v2/v1);     
+    half v4 = fabs(v2/v1);     
     printf("TENSOR = %f  | ------  MXM = %f  ----- | CUBLAS =  --------| RELATIVE = %.15f --------| \n", v1, v2, v4);
 
   }
